requirements.txt
pillow
lingua-language-detector
